added Definitions + solved missing text sections
The definitions include some of the more frequently used definitions for the requirements doc.
Each Section subsection and subsubsection must have a paragraph according to standard documentation. Therefore this was added. These are subject to change if necessary.
Also removed instances of Swedish social security number and replaced it with PN (sorry Henry)